algorithms: Añadidos algunos algoritmos y eliminados otros crones: Añadido un nuevo entrenador de modelos html: Refactor del js y el html

+++ algorithms/logisticregression/model/charge_model.py
@@ -10,7 +10,7 @@ xg_reg = pickle.load(open(name, "rb"))
 
 data = json.loads(sys.argv[-1])
 
-dataset = pd.DataFrame(data=data, dtype=float)
+dataset = pd.DataFrame(data=data, dtype=int)
 
 preds = xg_reg.predict(dataset[0:])
